04_Django/19-project/requirements.txt
django==3.2.0
djangorestframework-simplejwt==5.2.0
djangorestframework